# api/openapi.json
{
  "components": {
    "parameters": {
      "cat": {
        "description": "Keep only these categories, comma-separated or repeated. Unrecognised values fall back rather than erroring.",
        "in": "query",
        "name": "cat",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
...
      "id": {
        "description": "Narrow the board to an explicit set of service ids, comma-separated or repeated. Set by a finding whose subject has nothing else in common. Unrecognised values fall back rather than erroring.",
        "in": "query",
        "name": "id",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
...
      "period": {
        "description": "The window the figures and the ranking cover: `24h`, `7d`, `30d` or `90d` as shipped. Unrecognised values fall back rather than erroring.",
        "in": "query",
        "name": "period",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "q": {
        "description": "Free-text search over names and descriptions. Unrecognised values fall back rather than erroring.",
        "in": "query",
        "name": "q",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "region": {
        "description": "A region id. Only meaningful when the scope is sub-national. Unrecognised values fall back rather than erroring.",
        "in": "query",
        "name": "region",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "role": {
        "description": "Which side of the exchange: `issuer` or `requestor`. Unrecognised values fall back rather than erroring.",
        "in": "query",
        "name": "role",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "scope": {
        "description": "Which population is on view. The deployment's declared scopes; as shipped `national` or `state`. Unrecognised values fall back rather than erroring.",
        "in": "query",
        "name": "scope",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "signals": {
        "description": "Which set of findings the signals band shows: `attention` or `opportunity`. Unrecognised values fall back rather than erroring.",
        "in": "query",
        "name": "signals",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
...
      "status": {
        "description": "Keep only these statuses, comma-separated or repeated. Unrecognised values fall back rather than erroring.",
        "in": "query",
        "name": "status",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
...
    },
...
    "securitySchemes": {
      "ingestToken": {
        "description": "The value of the environment variable named by `push.tokenEnvVar` in `config/sources.yaml`, shipped as `DPI_INGEST_TOKEN`. Never a value from a configuration file: a token in a config file ends up in a git history.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Two ways to get data in, and they are alternatives rather than layers.\n\n**Push** — your collector calls `POST /api/v1/ingest` whenever it has something to say. It is the richer of the two: incidents and error buckets can only arrive this way, and every record is validated against the deployment's own taxonomy, so a category nothing declares is rejected rather than quietly becoming unreachable.\n\n**Pull** — the dashboard polls an endpoint you already have, on a timer, and maps its shape onto its own with a declarative mapping in `config/sources.yaml`. Nothing to build and nothing to schedule, but it can carry only what a mapping can express. `POST /api/v1/pull/preview` exists so that mapping can be worked out interactively rather than by editing a file and restarting.\n\nExactly one of the two is active, chosen by `driver:` in `config/sources.yaml`, so an operation this deployment does not register says so in its own description.\n\n**What the dashboard decides for itself.** Status is never reported, only derived: a collector sends observations and the dashboard evaluates them against the thresholds it publishes. `status`, `trends` and `rankMovement` are absent from the ingest message rather than ignored, so the contract states it plainly.",
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache-2.0"
    },
    "summary": "Feed a dashboard, read back what it holds, and try a pull mapping before committing to one.",
...
  },
...
  "openapi": "3.1.0",
  "paths": {
    "/": {
      "get": {
        "description": "Every reader selection rides in the query string, so a shared link reproduces exactly what the sender was looking at. Unrecognised values fall back to the deployment's defaults rather than erroring: a hand-edited URL or a bookmark that outlived a config change should produce a sensible page, not an empty one.",
        "operationId": "dashboard",
        "parameters": [
          {
            "$ref": "#/components/parameters/scope"
          },
          {
            "$ref": "#/components/parameters/role"
          },
          {
            "$ref": "#/components/parameters/id"
          },
          {
            "$ref": "#/components/parameters/signals"
          },
          {
            "$ref": "#/components/parameters/region"
          },
          {
            "$ref": "#/components/parameters/period"
          },
          {
            "$ref": "#/components/parameters/q"
          },
          {
            "$ref": "#/components/parameters/status"
          },
          {
            "$ref": "#/components/parameters/cat"
          },
          {
            "$ref": "#/components/parameters/sort"
          },
          {
            "$ref": "#/components/parameters/dir"
          },
          {
            "$ref": "#/components/parameters/lang"
          },
          {
            "$ref": "#/components/parameters/filters"
          },
          {
            "$ref": "#/components/parameters/theme"
          },
          {
            "$ref": "#/components/parameters/tab"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "The rendered markup."
          }
        },
        "summary": "The dashboard",
        "tags": [
          "Dashboard"
        ]
      }
    },
...
    "/assets/theme.css": {
      "get": {
        "description": "Built at startup from `config/theme.yaml`, never a file on disk. Served immutable with a fingerprint for an ETag, so a restyled deployment produces a different URL rather than a stale cache.",
        "operationId": "themeCSS",
        "responses": {
          "200": {
            "content": {
              "text/css": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "CSS custom properties."
          },
          "304": {
            "description": "Unchanged since the ETag the request carried."
          }
        },
        "summary": "The generated stylesheet",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/assets/{path}": {
      "get": {
        "description": "The stylesheet, the script, htmx, the fonts and the API console's renderer, all compiled into the binary. Served immutable for a year, which is safe because every link carries a `?v=` fingerprint over the whole embedded tree.",
        "operationId": "asset",
        "parameters": [
          {
            "description": "A path under `web/static`, e.g. `app.css`.",
            "in": "path",
            "name": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The asset."
          },
          "404": {
            "description": "No such asset."
          }
        },
        "summary": "A static asset",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/fragments/service/{id}": {
      "get": {
        "description": "The whole panel, for opening one.",
        "operationId": "drawerFragment",
        "parameters": [
          {
            "description": "A service id, as returned by `GET /api/v1/services`.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/scope"
          },
          {
            "$ref": "#/components/parameters/role"
          },
          {
            "$ref": "#/components/parameters/id"
          },
          {
            "$ref": "#/components/parameters/signals"
          },
          {
            "$ref": "#/components/parameters/region"
          },
          {
            "$ref": "#/components/parameters/period"
          },
          {
            "$ref": "#/components/parameters/q"
          },
          {
            "$ref": "#/components/parameters/status"
          },
          {
            "$ref": "#/components/parameters/cat"
          },
          {
            "$ref": "#/components/parameters/sort"
          },
          {
            "$ref": "#/components/parameters/dir"
          },
          {
            "$ref": "#/components/parameters/lang"
          },
          {
            "$ref": "#/components/parameters/filters"
          },
          {
            "$ref": "#/components/parameters/theme"
          },
          {
            "$ref": "#/components/parameters/tab"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "The rendered markup."
          }
        },
        "summary": "One service's drawer",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/fragments/service/{id}/pane": {
      "get": {
        "description": "The pane below the tab strip, plus the strip itself as an out-of-band swap. Everything above it is the same whichever tab is showing, so changing tab does not rebuild it.",
        "operationId": "drawerPaneFragment",
        "parameters": [
          {
            "description": "A service id, as returned by `GET /api/v1/services`.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/scope"
          },
          {
            "$ref": "#/components/parameters/role"
          },
          {
            "$ref": "#/components/parameters/id"
          },
          {
            "$ref": "#/components/parameters/signals"
          },
          {
            "$ref": "#/components/parameters/region"
          },
          {
            "$ref": "#/components/parameters/period"
          },
          {
            "$ref": "#/components/parameters/q"
          },
          {
            "$ref": "#/components/parameters/status"
          },
          {
            "$ref": "#/components/parameters/cat"
          },
          {
            "$ref": "#/components/parameters/sort"
          },
          {
            "$ref": "#/components/parameters/dir"
          },
          {
            "$ref": "#/components/parameters/lang"
          },
          {
            "$ref": "#/components/parameters/filters"
          },
          {
            "$ref": "#/components/parameters/theme"
          },
          {
            "$ref": "#/components/parameters/tab"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "The rendered markup."
          }
        },
        "summary": "One tab of an open drawer",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/fragments/{section}": {
      "get": {
        "description": "What an HTMX swap asks for. The section ids come from the deployment's `layout.yaml`; as shipped they are `verdict`, `signals` and `leaderboard`. The response carries an `HX-Push-Url` header naming the reader-facing address the fragment corresponds to.",
        "operationId": "sectionFragment",
        "parameters": [
          {
            "description": "A section id declared in `layout.yaml`.",
            "in": "path",
            "name": "section",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/scope"
          },
          {
            "$ref": "#/components/parameters/role"
          },
          {
            "$ref": "#/components/parameters/id"
          },
          {
            "$ref": "#/components/parameters/signals"
          },
          {
            "$ref": "#/components/parameters/region"
          },
          {
            "$ref": "#/components/parameters/period"
          },
          {
            "$ref": "#/components/parameters/q"
          },
          {
            "$ref": "#/components/parameters/status"
          },
          {
            "$ref": "#/components/parameters/cat"
          },
          {
            "$ref": "#/components/parameters/sort"
          },
          {
            "$ref": "#/components/parameters/dir"
          },
          {
            "$ref": "#/components/parameters/lang"
          },
          {
            "$ref": "#/components/parameters/filters"
          },
          {
            "$ref": "#/components/parameters/theme"
          },
          {
            "$ref": "#/components/parameters/tab"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "The rendered markup."
          }
        },
        "summary": "One band of the page",
        "tags": [
          "Dashboard"
        ]
      }
    },
...
    "/service/{id}": {
      "get": {
        "description": "The same page with the drawer open, so a service is a real address that can be bookmarked and sent.",
        "operationId": "servicePage",
        "parameters": [
          {
            "description": "A service id, as returned by `GET /api/v1/services`.",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/scope"
          },
          {
            "$ref": "#/components/parameters/role"
          },
          {
            "$ref": "#/components/parameters/id"
          },
          {
            "$ref": "#/components/parameters/signals"
          },
          {
            "$ref": "#/components/parameters/region"
          },
          {
            "$ref": "#/components/parameters/period"
          },
          {
            "$ref": "#/components/parameters/q"
          },
          {
            "$ref": "#/components/parameters/status"
          },
          {
            "$ref": "#/components/parameters/cat"
          },
          {
            "$ref": "#/components/parameters/sort"
          },
          {
            "$ref": "#/components/parameters/dir"
          },
          {
            "$ref": "#/components/parameters/lang"
          },
          {
            "$ref": "#/components/parameters/filters"
          },
          {
            "$ref": "#/components/parameters/theme"
          },
          {
            "$ref": "#/components/parameters/tab"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "The rendered markup."
          }
        },
        "summary": "The dashboard with one service open",
        "tags": [
          "Dashboard"
        ]
      }
    }
  },
  "servers": [
...
}